backend/requirements.txt
gunicorn==20.1.0
Django==3.2.25
djangorestframework==3.13.1
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
django-filter==21.1
django-cors-headers==3.13.0
drf-yasg==1.21.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
